Fix timestamps in exports from osm api db (#2365)
* Better timestamp handling in the OsmApiDbReader class. (Copied from HootApiDbReader, which works fine)

* Fix unit tests. Add ability to ignore source:datetime to MapComparator.

* Changed the way timestamps are read from the database, so they are not converted from binary to string then back to binary again. Now we just convert binary to binary.

Updated internal timstamp storage in nodes, ways, relations to be quint64, instead of uint32... so we get ahead of that pesky 2038 problem.

Updated documentation for the map-diff command.

* Add unit test for MapComparator

* Fix issues with timestamps & UTC
Convert some more timestamp calls to quint64

* Update HootApiDbReader with UTC handling

* Update copyright headers...

+++ hoot-core/src/main/cpp/hoot/core/elements/Element.h
@@ -22,7 +22,7 @@
  * This will properly maintain the copyright information. DigitalGlobe
  * copyrights will be updated automatically.
  *
- * @copyright Copyright (C) 2015, 2016, 2017 DigitalGlobe (http://www.digitalglobe.com/)
+ * @copyright Copyright (C) 2015, 2016, 2017, 2018 DigitalGlobe (http://www.digitalglobe.com/)
  */
 #ifndef ELEMENT_H
 #define ELEMENT_H
@@ -104,8 +104,8 @@ public:
   long getVersion() const { return _getElementData().getVersion(); }
   void setVersion(long version) { _getElementData().setVersion(version); }
 
-  unsigned int getTimestamp() const { return _getElementData().getTimestamp(); }
-  void setTimestamp(unsigned int timestamp) { _getElementData().setTimestamp(timestamp); }
+  quint64 getTimestamp() const { return _getElementData().getTimestamp(); }
+  void setTimestamp(quint64 timestamp) { _getElementData().setTimestamp(timestamp); }
 
   QString getUser() const { return _getElementData().getUser(); }
   void setUser(QString user) { _getElementData().setUser(user); }
